feat: 개인 보따리 물품 편집 API 호출 함수 추가

+++ android/Bottari/data/src/main/java/com/bottari/data/model/item/SaveBottariItemsRequest.kt
@@ -8,5 +8,5 @@ data class SaveBottariItemsRequest(
     @SerialName("createItemNames")
     val createItemNames: List<String>,
     @SerialName("deleteItemIds")
-    val deleteItemIds: List<Int>,
+    val deleteItemIds: List<Long>,
 )
